+++ Editor/Scripts/Drawers/FieldDisplayer.cs
@@ -10,10 +10,11 @@ namespace RealityProgrammer.OverseerInspector.Editors.Drawers {
 
             DrawAllChildsLayout();
 
-            BeginHandleReadonly();
-            BeginHandleFieldAssignCallback();
-            EditorGUILayout.PropertyField(AssociatedMember.Property);
-            EndHandleFieldAssignCallback();
+            if (BeginHandleReadonly()) {
+                BeginHandleFieldAssignCallback();
+                EditorGUILayout.PropertyField(AssociatedMember.Property);
+                EndHandleFieldAssignCallback();
+            }
             EndHandleReadonly();
         }
     }