+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.2.1] - 2021-12-06
+### Fixed
+- Field assign callback throws null when no callback are assigned
+
+### Changed
+- Changed how ReadonlyFieldAttribute works
+
 ## [1.2.0] - 2021-12-02
 ### Fixed
 - Inheritance reflection cache